Organizational

+++ src/vdp/vdp.h
@@ -1,27 +1,27 @@
 #ifndef SEGA_VDP_EMULATOR_H
 #define SEGA_VDP_EMULATOR_H
 
-#include "utilities.h"
+#include <cstdint>
 #include <functional>
 
 struct vdp {
 
     vdp();
 
-    uint8 reg[16];
-    uint8 mode;
+    uint8_t reg[16];
+    uint8_t mode;
 
-    uint8 vram[16 * 1024];
-    uint8 cram[64];
+    uint8_t vram[16 * 1024];
+    uint8_t cram[64];
 
     enum StatusFlags {
         VBlank              = 0b10000000,
         SpriteOverflow      = 0b01000000,
         SpriteCollision     = 0b00100000
     };
-    uint8 status;
+    uint8_t status;
 
-    int deviceType;
+    uint8_t deviceType;
 
     bool cycle();
     bool canSendInterrupt();
@@ -30,51 +30,51 @@ struct vdp {
     bool requestFrameInterrupt;
 
     /* Control port */
-    uint8 controlOffset;
-    uint16 controlWord;
+    uint8_t controlOffset;
+    uint16_t controlWord;
 
-    uint8 readControlPort();
-    void writeControlPort(uint8 data);
-    uint8 getControlCode();
-    uint16 getControlVRAMAddress();
+    uint8_t readControlPort();
+    void writeControlPort(uint8_t data);
+    uint8_t getControlCode();
+    uint16_t getControlVRAMAddress();
     void incrementControlVRAMAddress();
-    uint8 getControlRegisterIndex();
-    uint8 getControlRegisterData();
+    uint8_t getControlRegisterIndex();
+    uint8_t getControlRegisterData();
 
     /* Data port */
-    uint8 readBuffer;
+    uint8_t readBuffer;
 
-    uint8 readDataPort();
-    void writeDataPort(uint8 data);
+    uint8_t readDataPort();
+    void writeDataPort(uint8_t data);
 
     /* TV H/V scanners, active / inactive areas */
-    uint16 vCounter;
-    uint16 hCounter;
+    uint16_t vCounter;
+    uint16_t hCounter;
 
-    uint8 readHCounter();
-    uint8 readVCounter();
-    uint16 getActiveDisplayWidth();
-    uint16 getActiveDisplayHeight();
-    uint16 getHCounterLimit();
-    uint16 getVCounterLimit();
+    uint8_t readHCounter();
+    uint8_t readVCounter();
+    uint16_t getActiveDisplayWidth();
+    uint16_t getActiveDisplayHeight();
+    uint16_t getHCounterLimit();
+    uint16_t getVCounterLimit();
 
     /* TV screen sizes */
-    uint16 getScreenWidth();
-    uint16 getScreenHeight();
-    uint16 getScreenOffsetX();
-    uint16 getScreenOffsetY();
+    uint16_t getScreenWidth();
+    uint16_t getScreenHeight();
+    uint16_t getScreenOffsetX();
+    uint16_t getScreenOffsetY();
 
     /* VRAM addresses */
-    uint16 getNameTableBaseAddress();
-    uint16 getSpriteTableBaseAddress();
+    uint16_t getNameTableBaseAddress();
+    uint16_t getSpriteTableBaseAddress();
 
     int frameBuffer[256 * 240];
 
     void drawScanLine();
-    void drawTile(uint16 tileIndex, int x, int y, bool horizontalFlip=false, bool verticalFlip=false, bool spritePalette = false, bool doubleScale = false, bool tileWrap = false);
+    void drawTile(uint16_t tileIndex, int x, int y, bool horizontalFlip=false, bool verticalFlip=false, bool spritePalette = false, bool doubleScale = false, bool tileWrap = false);
     void drawTilemap(bool drawPriority);
     void drawSprites();
-    int getColor(uint8 paletteIndex);
+    int getColor(uint8_t paletteIndex);
 };
 
 #endif
